Add C# implementation of LIGO's "Efficient Algorithm for computing a Running Median", which computes a median over a moving window of size M over a data set of size N. The origial paper says that it is O(N * sqrt(M)) worst case vs O(N * (M * log (M))) average case for re-sorting every moving window of size M (worst case is O(N*(M^2))) . For M=20/200 this is 19x/108x difference.
Works in a script already, needs proper refactoring and tests. It is an online algorithm (its main loop updates the state incrementally) and could be easily turned into a Spreads cursor similar to SMA (that is the goal of porting it).

Source: https://dcc.ligo.org/public/0027/T030168/000/T030168-00.pdf