Fix CompressedArrayBinaryConverter serialization of IDelta types. Calculate DateTime deltas from previous values as a special and default case, but support old DT data that was not diffed.
A comment from code:
For DateTime we calculate delta not from the first but
from the previous value. This is a special case for the
fact that DT[] is usually increasing by a similar (regular) step
and the deltas are always positive, small and close to each other.
In contrast, Price/Decimal could fluctuate in a small range
and delta from previous could often change its sign, which
leads to a very different bits and significantly reduces
the Blosc shuffling benefits. For stationary time series
deltas from the first value are also stationary and their sign
changes less frequently that the sign of deltas from previous.