Implement `RetainableMemoryPool` support for small pinned array slices. Pool inherits `MemoryPool<T>` and slices inherit `Memory<T>`, could be used with pipelines. Should fix pinned fragmentation & OOM issue with our Json serialization using many small pinned buffers.
Add `ArrayMemorySliceBucket` and pool pinned array slices of 128kb slabs that are always in LOH.

Performance is very good (same code as in ConfigurableArrayPool from coreclr on hot pooled path). No GCHandle allocation is needed for `Memory.Pin` since slabs are pinned, 37MOPS for `ArrayMemorySlice` vs 44 MOPS for simple `ArrayMemory`. Not clear why the diff given the code is exactly the same and only `RetainableMemoryPool` is hit.

Slabs themselves are backed by pooled arrays and if `RetainableMemoryPool` usage is wrong and rented arrays are not returned (dropped or disposed) then slabs will return to the shared pool when all their slices are finalized. In the worst case of dropping 1M buffers in a loop the private set after test is 148MB and perf including waiting for all finalizers is:

 Case                |    MOPS |  Elapsed |   GC0 |   GC1 |   GC2 |  Memory
---------------------|--------:|---------:|------:|------:|------:|--------:
FullCycle            |    0.09 | 10681 ms |  29.0 |  22.0 |  16.0 | 13.204 MB

With pooling, the same test (100M count):

 Case                |    MOPS |  Elapsed |   GC0 |   GC1 |   GC2 |  Memory
---------------------|--------:|---------:|------:|------:|------:|--------:
FullCycle            |   36.38 |  2749 ms |   0.0 |   0.0 |   0.0 | 0.133 MB (NB: single slab 128kb)

Should watch for new GC API for large array allocation to replace current implementation for large buffers:

https://github.com/dotnet/corefx/issues/31787
https://github.com/dotnet/coreclr/issues/19936