SizeOf signature returns (buffer,actualFormat) because given format if preferred best-effort one.
We fallback to JSON or skip compression of small values. Previously, SizeOf wrote header to payload,
but now it returns only the size of value and without returning actual format used we lose that info
and Write has no way to recover it. Pass tuple as `in` parameter, overhead is visible.

Performance for the fixed-size case is not affected.

Complex signature is not a problem and supposed usage is `out var` goes directly to `in`:

```
var sizeOf = BinarySerializer.SizeOf(in val, out var payload, preferredFormat);
// allocate buffer of at least DataTypeHeader.Size + sizeOf
var written = BinarySerializer.Write(in val, buffer, in payload, preferredFormat);
if (DataTypeHeader.Size + sizeOf != written)
{
    Assert.Fail("DataTypeHeader.Size + sizeOf != written");
}
```