Nested tuple binary serialization works at 6 Mops, just 2x faster than JSON.
Should have been at least 5x faster. The problem is array pool - it's actually not that fast at 30 Mops.
Every nested operation requires Retain/Dispose which is around 1/3 of total job.

We should learn from e.g. RIO and manage temp buffer outside BS.
Also JsonWriter/Reader (or Binary Writer/Reader) is a simpler pattern.
We should keep SizeOf/Write separation because we do need to know
the size before writing to DS. But when the size is not fixed then all job
is done inside SizeOf with writer pattern.

SizeOf could return `out int payloadWritten`, when it is not zero then
BinaryWriter added that length of payload to the writer. For the nested case
we could call inner SizeOf and if it's fixed write to the same buffer directly.
When all nested items are processed we have complete outer payload.