Refactor VectorStorage
It's only RetainableMemory<T> stored as IRefCounted + materialized Vec (RM has it as property but creates on demand).

Length is underlying Vec length and means capacity. Actual number of elements should be stored in DataBlock. Stride only matters for matrices and should be implemented in Vector/Slice that represents a matrix row/column.

VectorStorage could be made a struct later if it is always owned by a disposable+finalizable container.

Also misc changes and enable C# 8.0 nullable.