Move new containers to the top project, deprecate old Series'2
New containers are temporary in Spreads.X namespace to avoid conflicts. Old containers will live for a while in Spreads.Deprecated NS.
Then SortedMap and Spreads.Collections will live on their own and won't be a dependency of the top Spreads package. This is only to support legacy code for a while, but SortedMap should be easily replaced by just Series and therefore will be there just to monitor performance regressions.