Use BenchmarkDotNet from unit tests in-process.
Add a base test fixture that allows to easily maintain benchmarks as unit tests and run in-process.

This makes the workflow close to Spreads.Benchmark util class. But BDN is not a silver bullet and
requires a lot of turning to get meaningful numbers when we count cycles, not millis. The current
util is quite precise when used from NoInlining+AggressiveOpt method and JIT optimizations
are warmed up.